Сделал правки. Вынес все функции в отдельные файлы. Подключил библиотеку.
