+++ Task01/Header.h
@@ -1,0 +1,24 @@
+#pragma once
+
+int Power(int a);
+
+double PowerMH(double a);
+
+double IndexMass(double height, double weight);
+
+int MaxTwo(int a, int b);
+
+int MaxFour(int a, int b, int c, int d);
+
+void Swap(int* k, int* f);
+
+void SwapNotThree(int* k, int* f);
+
+void QuadraticEquation(double a, double b, double c);
+
+int Discriminant(double a, double b, double c);
+
+void Task01();
+void Task02();
+void Task03();
+void Task04();
