+++ Task01/IndexMass.c
@@ -1,0 +1,15 @@
+#include "Header.h"
+
+double IndexMass(double height, double weight)
+{
+	double res = 0;
+	//	I=m/(h*h)
+	if (height > 3)
+	{
+		return weight / Power(height);
+	}
+	else if (height < 3)
+	{
+		return weight / PowerMH(height);
+	}
+}
